TLS自适应测试代码修改
