+++ gmtls/websvr/img/autoswitchlogic.png
