--- gmtls/websvrcase/certs/RSA_CA_KEY.pem
@@ -1,27 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEpAIBAAKCAQEAzImNwnLsiwtQYdcHkqbm24ui+w+3RUdaPIsp4Qv201VL2zPc
-6UzFP+eLfUImaPmiEOatsEcKaVtdHgS9limBTaOlvS3EqqA1EoorsfEV7VPBDxnC
-R2zkD3Mpu0WdUPErRm2rlz1YG77j+690PUDFa0Eg6zdhJPg/PJgqgsKJEmkXtCIu
-le0p7cdQ3voHiwJN64B1HWp3XS5329jxBTqZwaFkrVCR+R58QzGgcajdz+yTJmZk
-x63FK9MKHSiU5RxK+gywYz+tiuLtWKIeRw/PkUVFRX2+2lx3rYEXj+cwJJlIbWNN
-J9yixd4OFBXDcX9S7bJ3/1WtViuj/rUXhRvo7wIDAQABAoIBAEJngiOYJh26TZ2G
-L0o4b4jlm8NAxfJzCOPRFoae2MKh1Z/UBaukpgFqTgobmNbJ/QjZq5goaEPuSpv9
-VAPVc8peD+gMXOY8Xx6mxyrdJ45Ge85IQZ/oxV4SuIC5SCrIMZGK94OBggJ4BhJn
-Uyizx6Rbn7Fd6GcVRYqJqHuNZu0XprculGZ58CA//EJejDm0QZc8lqFlis5WCJ/G
-GMV1xMNx/QhU7ObbzVGfIUeDyYCsO7F3SxWbo9cyToQi/14MeI/kLhSp9CPv2PSL
-ezkSazVrRxNqOJBzlTncA+z0p+RLdfvkNB82CAtJdnN94VF5H1/dePkn4HhT9hz5
-vP55HZECgYEA8QNqw7qwH6jfRpaAcJ8DP8MPzZVPF7rliK4buQDcE/5LzIP/5BhK
-dCT+vKYHcPGF3IqsQ6rZDZ4aUvlzdXi8yolz6NNDgEGYYiQH0qNFY1DnR+YnqtuS
-o1+LmZXJkuRk/6w2/Fd+EDK+7SLLQLKDGb+bmReDvW3vye2DbN/mSicCgYEA2UF9
-wRA45+2gaK4cEkrovq6j6U8+IYPuQFXVFVKhILGN69/l1ZLamVMGX5tMUwfwJdh8
-2Ewqz5tpdBrrCwdnhQHugZl9YFzLej98eLREOqiOcqAvFVjFsbtfe4jwFn7ciHuO
-Fe//xWrI9dVQLhL/cAuJMtTlWVCnAGCIhjj+j/kCgYEAsSVb0eVkEsEi61UvrpeR
-4cnT0b2O1thOYbnnMoHJQQDGYdobuWGJR/jxK2O1UHSVeVaL62imBPZSTu5+yLcn
-vI5vHZA6qwiLoobtIQtUsv0CIbyvJ0wgaeiOSSLpeAhbm/jOqYb9Z5aE8vOCBVmw
-Nrq4+TtpX1Bw0TC8Ssf/Ej0CgYAnbA7ZnYv16dSVqiJF33nkP+RWTZ7EjiAJZInJ
-2LiMKivKhVI7Mu6fq8UYTCHJX0NgWqwsw4VxLo4wF0JXOfy60eLpeWSau7vgbuyz
-KmLHEhbnDdH9WaJl4c4tqGeIbXSu/1yE9ou+0uWsVgy6j4/yUOWQZ7lGP7BODTY4
-xNbMYQKBgQCn3AhkE2di4XHdKZOr3aSW6yJovZl+7bF1/xiMuOh5Noz7h6RwVGa4
-2ZJnRyyV3oeOIftkBNls42LPpqo1O2Y7vOdxLtdh9DLJbVAUgobvIufpJ6VSUhmP
-THNdL0/ciJ3vP9zLtcVVYR/2IyI+lMbnyhD8BNvFzbblK1NUSJiStQ==
------END PRIVATE KEY-----
