--- gmtls/websvrcase/certs/SM2_CA_KEY.pem
@@ -1,6 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIGTAgEAMBMGByqGSM49AgEGCCqBHM9VAYItBHkwdwIBAQQg8/o/oC+JT0Nv2T0K
-AdZ8OiH9YGVIGsnx0xfDK+lXzpagCgYIKoEcz1UBgi2hRANCAAS6OFCNygoH2f4O
-nPbV570rydtVcTvAUYOo/Mk9dcKsvEwDu+WZ2Lw8Ef4PCcqgm6B6+qo86x4AKXjm
-pTzXvXf9
------END PRIVATE KEY-----
