--- gmtls/websvrcase/certs/rsa_auth_key.pem
@@ -1,27 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEogIBAAKCAQEA6TZW838H5RS+gN2bFVlMDplAf8Fl8PIAKPeyxUJs0bBmWWeN
-tvkagsED7dqlDoY1jmnmXNmgAj4JH3TuD0GOTh/2yj03P3vIO7iQbMU99ZASpYBv
-dH8cD8ildLItZwJdytg9unxwVtPEwi0MM71GnAnji7+el2TlNqCJQkv9R0feEO/0
-hzxK+dwMHjlI/fto3NWs2MqLo95lsM1KHbBW6Zbku/D6soUsMohFWcgDOKxCtWt6
-cMQNcdrBXjYncUSNDarxkjBUbMdYF+Lg+fqTfJ30YCDUcwLBqmSg801y5sAijTti
-BwckxJsAogEban9EoGzc+xWsHGJ7vuLLmNNGnQIDAQABAoIBAF2ev9cuJrgy1Ksc
-yzYNy/eMVBU1STwpmiDYFv8TRGYFP4EijgRnCdp1q6Mfg+VJwVQzdRPPFWnxB26G
-XU7ojT0VndfgLT6h7Wb0F55irBXZSXMBFT0aDeTUMn+fS5SvoYFdv6IOdHaWusUM
-mnpmwGLBoYQnX339NfrxBUkFAl9LWUrw3Y+ON9R6uEeSMntD84L4R9R+84S5BeTr
-6zb/XoT4YvD6RS8HOB7XBCGlLTwds+a4zyQnuCCHbKIES9lxItF0QbprG5J+Q7gp
-lebi/zXLWbUg22S3mzTmP3R8l0LWwKhyPnBh8onNf/Nroa68r9FlEFSglj/5MAvJ
-pAD6/80CgYEA975QPGmW6mxBxUr3OVZTKmTVGHonEb/qVr489zgnruWwbY7Y344V
-qw2nY+EiT2ftGsZmb/LZ3fEaSdmKYb1rDS9Jslfd3E1yUygFFtwqiGqWEp8LDDvU
-UgqDvMdoDaGIKcl0EyeWOjCi073B2MSC9prKY9MuxHpOfZ7f4aN7GbsCgYEA8PwM
-zZZCAWRFlasxYe59uU63vSmDGkMcy0W494IrfcqPLIoin4Xkdrjb1LVI6dYb27/W
-2gFMKMlWCHCvaayFzgmwjIK8+r+nUWAuxCWljxOaB3wIKf3q1sNVAITfFTuV1k6d
-Svl55fpYAPqDu6JLSqoIZK+Vw2amTOnJSrZoT4cCgYBuZ5uuMIgW1EPDPpNKNBQ/
-8kuc6uL6YAJ4wbK9oYLdyHKFgIFZryIs5prQQgZeSLw7kGsaPKmm5MmIW6/1uso4
-R/iZx9qRUNpAALez5k32GTnj9oif3RHTtefbeyn92RyrMqHN33htWjtWDVyT049h
-gr3rGrPez3qnNOQk1AwJ/QKBgGeOuOXEE1tLvsdhiNgfx7FRTY4hf7BSdNY5gRID
-wzLMUBiAx/4raJFIQVe4C7xJATG+z9JhSlwgoTMBUbAkrl9rNI49RH6wy3Mh1Rb/
-YpUIhGGBAHUjI77ks+uELeWO9IXixJ9KdALLQPQys1d3l8lfJjo2ZjJRM1fMA29y
-0YKTAoGAA7adLMc8uePn6Dclw/MQ291t3JWYKOuZ5bM4BfwMf+Xr5XHuIylNaxgJ
-JyIkZcrnEpIjKwtwCBUf3fJ2I39d7Euhdss7yS+YVrvuQSINtoKRmZ45PKP0MKNf
-+YyctbsNPGmbzK93WBijRefdAIvyGG29FlXwnKP9A0tBXM7yx1o=
------END PRIVATE KEY-----
