--- gmtls/websvrcase/certs/rsa_sign_key.pem
@@ -1,27 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEowIBAAKCAQEAvxRqNv421N5OVGvjJApIHFL3UCTI4NncP842tM5x4hlJcn4N
-sFr/VuW164ERgQG0nhikojIHjyuN8aQRsKFjNE3agDsUDqqUDsAoMUobLbC8/t0d
-ms9Hc2t8LlRpc/T5/5mklAa60Ls8g8KGOl4UGBxEX27zMx0NisntNoXSHQRm4AO1
-bVCVAFpb43p7SJJ1CvgtxCijs9F4KRglWrnfVNLdkCt86UUzFpJSXCFx2FKu1FBQ
-RKu4J9xj280XuSw6dbIGIHps9XGkYFJ00S9NEfIyqlFKrx3Tlo62JPXOLzZjWXlS
-5zaVjVzPP0mN0S1zoUtZ3KySt91RSbGT4waiywIDAQABAoIBAHON1AH30I8A9ZvG
-NvHnqd4/KkL0S+r2tWR8v8auBl72bd1ob7zT+xDBEDggsMY0ktb9DirsWHfznf6g
-+vW3U7gg0N+1WpbX/3vAJ05Pg4opFNKIQ8wISCx7tviJodA3OSyi7M18GGP8p0cD
-T4tis4I1LgAMp9SkYDkrjmHnob5biilhdAGseJWjyKMxBc1sqMQVy+Gia49guU0J
-fnEOu1jT6P/ZB7P7wyiFEVJKYvU/iMQ2rCiBlBKmAa4ZuV/rMLs7/twwiD1EtMNc
-bpwoNeEI7XC+HCwpYNecscIxhs3zcjqasc93B7Sn0MHhdxz2aU6TH2BSA+S1zfzm
-+fcgS1kCgYEA/Tl3WERs2VjNW01c5yQ5fz3yKhqHgBZ9KODasLdX4/4b18ZRiQdI
-VRCMSjhp0IryYgkH78X6fKIdzA276MNpAK3mcCu6dI6d8ige55A4cwyS3P23b2Ft
-VB9TAHUJ+5F3fXHAjXYBzMsDNCkZqedPRj18+GMyHfMV05eqzEzR/p8CgYEAwSyS
-9NDir1lQVcjDy4IF8cjsuFT3R1KbUy29MOCBcTd4RaB/vTEs4nCZlpb9wBBJqQBu
-5CCH5wOxc+WvoEicpzrcwqWBTmUngulVHumub8VeahTHZsbPXDlc6hXg+kXimo0S
-rNGHGz76aBEgxpD8KnoRiBXL5KUXkYVSeFtT6FUCgYBlwRTbDEvtHskOekztuNm9
-7seD8bsnVmIe7hQS1OnlfPP1RgT3mkyPmiG1pWun2F3D1TsyN40LTh3JJG+w/e0Q
-z9B8nKXFNpXze/nHNWXpnVDaqr1ubWkAV8IDugQz6eVpLC+i6Juajq2P5UAhzQhH
-AmhYa4v1CvxnVVfV1ry/owKBgHJrFnhPqZU3D7BgBCjsUN0Pp7bfDMCAG2Q/OUpU
-EBehRWnoUQEcuXs+OprJwG8ctqSVjxksc5LPyzHcoBQOLkuXP99C7L4kiCywTCzN
-OPcXLn8i46DqrTPw/ing36sd7cShPqiCSsZGnUc6rfEvO01llDDIyEYby0P0dg63
-+ZEpAoGBAPQQY2ojWp8smOpQhtkqPjeDb2rjSGoniBz3mapFv4WuPmpQhRKh/dcu
-gCIHmXOAEz8nLCKWF+CsKpvLIuO+nyT2XadPuz3nAJM6VwVZ+wrQUEI6tP4DcfqB
-SDcVFeEKFw+nZGMFfTm3g5/eFraRjSBK5a5immGBMJCQQanniN7B
------END PRIVATE KEY-----
