--- gmtls/websvrcase/certs/sm2_auth_key.pem
@@ -1,6 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIGTAgEAMBMGByqGSM49AgEGCCqBHM9VAYItBHkwdwIBAQQgqSvMdgsT9VlcC7fF
-uewfAWp4eNoLT3zhTZisHsZ2WS+gCgYIKoEcz1UBgi2hRANCAAQVZNo44m183rOc
-7tZ4qHL0tnYSL1RXnzdXCiUfH36Kku1yJ9DKKm6f5BtOPUzWx3YEy0UrrzXucie1
-PdwKFtWm
------END PRIVATE KEY-----
