--- gmtls/websvrcase/certs/sm2_sign_key.pem
@@ -1,6 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIGTAgEAMBMGByqGSM49AgEGCCqBHM9VAYItBHkwdwIBAQQguBgnxgDxfYC9AM3v
-C2oQ7W54T7XLS3Dfoiv/8bQQXZ2gCgYIKoEcz1UBgi2hRANCAAQGNFBLmSXwgypy
-AFln6qlgaV+a6H1RaTBrpF12ciuY+bna9ewBXbAdM2wYPBxq8ifZfl/TngvX8TAq
-rNwx8prv
------END PRIVATE KEY-----
